+++ packages/kiarina-lib-firebase/src/kiarina/lib/firebase/_instances/token_manager_registry.py
@@ -31,6 +31,7 @@ def _factory(settings_key: str, settings: FirebaseSettings) -> TokenManager:
     return TokenManager(
         api_key=settings.api_key.get_secret_value(),
         token_store=FileTokenStore(settings.token_data_file_path),
+        uid=settings.uid,
     )